+++ R/facto_pca.R
@@ -29,9 +29,11 @@ facto_pca <- function(X, ncp = 5, scale.unit = TRUE, ind_sup = NULL, quanti_sup 
     weighted_col <- rep(1, ncol(X_active))
   }
 
-  X_active_scaled <- standardize(X_active, scale = scale.unit)
+  ncp <- min(ncp, ncol(X_active))
 
-  eigs <- get_weighted_eigen(X_active_scaled, weighted_col = weighted_col)
+  X_active_scaled <- pca_standardize(X_active, scale = scale.unit)
+
+  eigs <- pca_weighted_eigen(X_active_scaled, weighted_col = weighted_col)
   eigvalues <- eigs$values
 
   df_eigs <- data.frame(
@@ -43,6 +45,8 @@ facto_pca <- function(X, ncp = 5, scale.unit = TRUE, ind_sup = NULL, quanti_sup 
   rownames(df_eigs) <- paste0("comp ", 1:nrow(df_eigs))
 
   ind_coords <- pca_ind_coords(eigs)
+  rownames(ind_coords) <- rownames(X_active)
+
   ind_cos2 <- pca_ind_cos2(ind_coords, weighted_col = rep(1, ncol(ind_coords)))
   ind_contrib <- pca_ind_contrib(ind_coords, eigs, weighted_row = rep(1, nrow(ind_coords)) / nrow(ind_coords))
 
@@ -67,35 +71,47 @@ facto_pca <- function(X, ncp = 5, scale.unit = TRUE, ind_sup = NULL, quanti_sup 
   weights <- rep(1, nrow(X_active)) / nrow(X_active)
   weighted_col <- rep(1, ncol(X_active))
 
+  lst_eigs <- eigs
+  lst_eigs[["values"]] <- sqrt(lst_eigs[["values"]])
+
+  center <- colMeans(X_active)
+  std <- sqrt(as.vector(crossprod(weights, as.matrix(X_active^2)) - center^2))
+
   res_pca <- list(
     eig = df_eigs,
     var = lst_var,
     ind = lst_ind,
+    svd = lst_eigs,
     call = list(
       row.w = weights,
       col.w = weighted_col,
       scale.unit = scale.unit,
-      ncp = ncp
+      ncp = ncp,
+      centre = as.vector(center),
+      ecart.type = std,
+      X = X,
+      row.w.init = rep(1, nrow(X_active)),
+      call = match.call()
     )
   )
 
   if (!is.null(ind_sup)) {
-    center <- colMeans(X_active)
-    std <- sqrt(as.vector(crossprod(weights, as.matrix(X_active^2)) - center^2))
-
     if (!is.null(quanti_sup)) {
-      X_sup <- X[ind_sup, -quanti_sup]
+      X_sup <- X[ind_sup, -quanti_sup, drop = FALSE]
     } else {
       X_sup <- X[ind_sup, ]
     }
 
-    X_sup_scaled <- (X_sup - center) / std
-    ind_sup_coords <- as.matrix(X_sup_scaled) %*% eigs$vectors
+    X_sup_scaled <- t(t(t(t(X_sup) - center)) / std)
+    ind_sup_coords <- as.data.frame(as.matrix(X_sup_scaled) %*% eigs$vectors)
+    rownames(ind_sup_coords) <- rownames(X_sup)
 
     res_pca$ind.sup <- list(
       coord = ind_sup_coords[, 1:ncp],
       cos2 = pca_ind_cos2(ind_sup_coords)[, 1:ncp]
     )
+
+    res_pca$call$ind.sup <- ind_sup
   }
 
   if (!is.null(quanti_sup)) {
@@ -105,14 +121,17 @@ facto_pca <- function(X, ncp = 5, scale.unit = TRUE, ind_sup = NULL, quanti_sup 
       X_sup <- X[, quanti_sup, drop = FALSE]
     }
 
-    X_sup_scaled <- standardize(X_sup, scale = scale.unit)
-    var_sup_coords <- t(X_sup_scaled * weights) %*% eigs$U
+    X_sup_scaled <- pca_standardize(X_sup, scale = scale.unit)
+    var_sup_coords <- as.data.frame(t(X_sup_scaled * weights) %*% eigs$U)
+    names(var_sup_coords) <- paste0("Dim.", 1:ncp)
 
-    res_pca$var.sup <- list(
-      coord = var_sup_coords,
-      cor = var_sup_coords,
-      cos2 = pca_var_cos2(var_sup_coords)
+    res_pca$quanti.sup <- list(
+      coord = var_sup_coords[, 1:ncp],
+      cor = var_sup_coords[, 1:ncp],
+      cos2 = pca_var_cos2(var_sup_coords)[, 1:ncp]
     )
+
+    res_pca$call$quanti.sup <- X_sup
   }
 
   class(res_pca) <- c("PCA", "list")