+++ R/pca.R
@@ -2,14 +2,14 @@
 #'
 #' Return principal component for individuals
 #'
-#' @param eigs eigs computed by \code{get_eigen} or \code{get_weighted_eigen}
+#' @param eigs eigs computed by \code{pca_eigen} or \code{pca_weighted_eigen}
 #'
 #' @examples
 #' library(FactoMineR2)
 #'
 #' iris[, -5] |>
-#'   standardize_norm() |>
-#'   get_weighted_eigen() |>
+#'   pca_standardize_norm() |>
+#'   pca_weighted_eigen() |>
 #'   pca_ind_coords() |>
 #'   head()
 #' @export
@@ -21,7 +21,7 @@ pca_ind_coords <- function(eigs) {
 
 #' Compute individual squared cosines
 #'
-#' Return indivdual squared cosines for each principal component
+#' Return individual squared cosines for each principal component
 #'
 #' @param ind_coords individual coordinates
 #' @param weighted_col column weights
@@ -30,8 +30,8 @@ pca_ind_coords <- function(eigs) {
 #' library(FactoMineR2)
 #'
 #' iris[, -5] |>
-#'   standardize_norm() |>
-#'   get_weighted_eigen() |>
+#'   pca_standardize_norm() |>
+#'   pca_weighted_eigen() |>
 #'   pca_ind_coords() |>
 #'   pca_ind_cos2() |>
 #'   head()
@@ -44,10 +44,10 @@ pca_ind_cos2 <- function(ind_coords, weighted_col = rep(1, ncol(ind_coords))) {
 
 #' Compute individual contributions
 #'
-#' Return indivdual contributions for each principal component
+#' Return individual contributions for each principal component
 #'
 #' @param ind_coords individual coordinates
-#' @param eigs eigs computed by \code{get_eigen} or \code{get_weighted_eigen}
+#' @param eigs eigs computed by \code{pca_eigen} or \code{pca_weighted_eigen}
 #' @param weighted_row row weights
 #'
 #' @details
@@ -58,8 +58,8 @@ pca_ind_cos2 <- function(ind_coords, weighted_col = rep(1, ncol(ind_coords))) {
 #' library(FactoMineR2)
 #'
 #' eigs <- iris[, -5] |>
-#'   standardize_norm() |>
-#'   get_weighted_eigen()
+#'   pca_standardize_norm() |>
+#'   pca_weighted_eigen()
 #'
 #' eigs |>
 #'   pca_ind_coords() |>
@@ -76,14 +76,14 @@ pca_ind_contrib <- function(ind_coords, eigs, weighted_row = rep(1, nrow(ind_coo
 #'
 #' Return variable coordinates
 #'
-#' @param eigs eigs computed by \code{get_eigen} or \code{get_weighted_eigen}
+#' @param eigs eigs computed by \code{pca_eigen} or \code{pca_weighted_eigen}
 #'
 #' @examples
 #' library(FactoMineR2)
 #'
 #' iris[, -5] |>
-#'   standardize_norm() |>
-#'   get_weighted_eigen() |>
+#'   pca_standardize_norm() |>
+#'   pca_weighted_eigen() |>
 #'   pca_var_coords() |>
 #'   head()
 #' @export
@@ -97,19 +97,19 @@ pca_var_coords <- function(eigs) {
 #'
 #' Return variable correlation
 #'
-#' @param eigs eigs computed by \code{get_eigen} or \code{get_weighted_eigen}
+#' @param eigs eigs computed by \code{pca_eigen} or \code{pca_weighted_eigen}
 #'
 #' @examples
 #' library(FactoMineR2)
 #'
 #' iris[, -5] |>
-#'   standardize_norm() |>
-#'   get_weighted_eigen() |>
+#'   pca_standardize_norm() |>
+#'   pca_weighted_eigen() |>
 #'   pca_var_cor() |>
 #'   head()
 #' @export
 pca_var_cor <- function(eigs) {
-  var_cor <- -1 * (eigs[["vectors"]] %*% diag(sqrt(eigs[["values"]])))
+  var_cor <- (eigs[["vectors"]] %*% diag(sqrt(eigs[["values"]])))
   colnames(var_cor) <- paste0("Dim.", 1:ncol(var_cor))
   return(var_cor)
 }
@@ -125,8 +125,8 @@ pca_var_cor <- function(eigs) {
 #' library(FactoMineR2)
 #'
 #' iris[, -5] |>
-#'   standardize_norm() |>
-#'   get_weighted_eigen() |>
+#'   pca_standardize_norm() |>
+#'   pca_weighted_eigen() |>
 #'   pca_var_coords() |>
 #'   pca_var_cos2() |>
 #'   head()
@@ -141,15 +141,15 @@ pca_var_cos2 <- function(var_coords) {
 #' Return variable contributions
 #'
 #' @param var_cos2 variable coordinates
-#' @param eigs eigs computed by \code{get_eigen} or \code{get_weighted_eigen}
+#' @param eigs eigs computed by \code{pca_eigen} or \code{pca_weighted_eigen}
 #' @param weighted_col column weights
 #'
 #' @examples
 #' library(FactoMineR2)
 #'
 #' eigs <- iris[, -5] |>
-#'   standardize_norm() |>
-#'   get_weighted_eigen()
+#'   pca_standardize_norm() |>
+#'   pca_weighted_eigen()
 #'
 #' eigs |>
 #'   pca_var_coords() |>